Separate rpc implementation by runtime
